# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "元数据驱动的企业级应用框架",
        "title": "Sky-Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:9090",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/cloud/files": {
            "get": {
                "description": "列出指定文件夹下的所有文件（支持分页）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "列出文件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件夹ID（0或空表示根目录）",
                        "name": "folderId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码（默认1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小（默认20）",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ListFilesResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/cloud/files/upload": {
            "post": {
                "description": "上传文件到指定文件夹",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "文件夹ID（0或空表示根目录）",
                        "name": "folderId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.CloudFile"
                        }
                    }
                }
            }
        },
        "/api/v1/cloud/files/{id}": {
            "delete": {
                "description": "删除指定文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "删除文件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/cloud/files/{id}/download": {
            "get": {
                "description": "下载指定文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "下载文件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/v1/cloud/files/{id}/move": {
            "put": {
                "description": "移动文件到指定文件夹",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "移动文件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "移动请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MoveFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/cloud/files/{id}/rename": {
            "put": {
                "description": "重命名指定文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "重命名文件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "重命名请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RenameFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/cloud/folders": {
            "get": {
                "description": "列出指定父文件夹下的所有子文件夹",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "列出文件夹",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "父文件夹ID（0或空表示根目录）",
                        "name": "parentId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.CloudFolder"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "在指定父文件夹下创建新文件夹",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "创建文件夹",
                "parameters": [
                    {
                        "description": "创建文件夹请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateFolderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.CloudFolder"
                        }
                    }
                }
            }
        },
...
        "/api/v1/cloud/folders/{id}": {
            "delete": {
                "description": "删除指定文件夹（包括其中的所有文件和子文件夹）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "删除文件夹",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件夹ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/cloud/folders/{id}/rename": {
            "put": {
                "description": "重命名指定文件夹",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "重命名文件夹",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件夹ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "重命名请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RenameFolderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/cloud/shares": {
            "post": {
                "description": "创建文件或文件夹分享链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "创建文件分享",
                "parameters": [
                    {
                        "description": "创建分享请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateShareRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.CloudShare"
                        }
                    }
                }
            }
        },
        "/api/v1/cloud/shares/{code}": {
            "get": {
                "description": "根据分享码获取分享信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "获取分享信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分享码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "访问密码（如果需要）",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cloud.ShareInfo"
                        }
                    }
                }
            }
        },
        "/api/v1/cloud/shares/{code}/access": {
            "post": {
                "description": "访问分享文件（需要密码验证）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "访问分享文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分享码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "访问请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AccessShareRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.CloudShare"
                        }
                    }
                }
            }
        },
        "/api/v1/cloud/shares/{id}": {
            "delete": {
                "description": "取消指定的文件分享",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cloud"
                ],
                "summary": "取消文件分享",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分享ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cloud.ShareInfo": {
            "type": "object",
            "properties": {
                "file": {
                    "$ref": "#/definitions/entity.CloudFile"
                },
                "folder": {
                    "$ref": "#/definitions/entity.CloudFolder"
                },
                "resourceType": {
                    "type": "string"
                },
                "share": {
                    "$ref": "#/definitions/entity.CloudShare"
                },
                "sharer": {
                    "description": "分享者名称",
                    "type": "string"
                }
            }
        },
...
        "handler.CreateFolderRequest": {
            "type": "object",
            "required": [
                "folderName"
            ],
            "properties": {
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "folderName": {
                    "description": "文件夹名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "父文件夹ID（nil表示根目录）",
                    "type": "integer"
                }
            }
        },
...
        "handler.ListFilesResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "handler.MoveFileRequest": {
            "type": "object",
            "properties": {
                "targetFolderId": {
                    "description": "目标文件夹ID",
                    "type": "integer"
                }
            }
        },
...
        "handler.RenameFileRequest": {
            "type": "object",
            "required": [
                "newName"
            ],
            "properties": {
                "newName": {
                    "description": "新名称",
                    "type": "string"
                }
            }
        },
        "handler.RenameFolderRequest": {
            "type": "object",
            "required": [
                "newName"
            ],
            "properties": {
                "newName": {
                    "description": "新名称",
                    "type": "string"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}